requirements.txt
fastapi
uvicorn
openai
slack_sdk
python-dotenv
slack_bolt 
aiohttp
httpx
